requirements.txt
catboost==1.2.8
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
graphviz==0.21
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.7
narwhals==2.13.0
numba==0.62.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pynndescent==0.5.13
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
shap==0.50.0
six==1.17.0
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
xgboost==3.1.2
